requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
python-multipart==0.0.9
towhee==1.1.3
faiss-cpu==1.7.4
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.1
transformers==4.38.2
sentencepiece==0.2.0
protobuf==4.25.3 
timm
opencv-python
